requirements.txt
google-generativeai
python-dotenv
Flask
flask-cors